Changes to README.md
