+++ README.md
@@ -1,16 +1,19 @@
-# Nailyzer
-##Overview
+# Nailyzer #
+##Overview##
 The project aims to detect symptoms of various diseases with use of image processing techniques.
 The software takes digital image as an input from user on which analysis is carried out. The software is trained with a dataset of nail images of about 279 individuals. Our algoithm extracts features based on color textures and irregularities in input image and outputs the probable cause of the same.
 This software is useful as it is one-of-its-kind and it can be used as a Self-diagnostic tool without the help of a medical professional.
 
-##What we have done
+##What we have done ##
 The users are required to upload images of their finger nails to the webserver. The webserver uses this uploaded files as input to the disease detection software.
 The software uses digital image processing algorithms to analyse the images of nails of users. Machine Learning is used to further classify the dataset image on the basis of texture and color of the nails.
  
 
-##Future Prospects
+##Future Prospects##
 As a nail is unique to every person, there are many body organs such as eyes and tongue whose images can be used to determine symptons of even a fatal disease in its early stages.
+
+##Requirements ##
+1) Octave or Matlab
